renamed units ui to wavelengths ui
