+++ cubeviz/layout.py
@@ -25,7 +25,7 @@ from .image_viewer import CubevizImageViewer
 from .controls.slice import SliceController
 from .controls.overlay import OverlayController
 from .controls.units import UnitController
-from .controls.units_ui import WavelengthUI
+from .controls.wavelengths_ui import WavelengthUI
 from .tools import arithmetic_gui, moment_maps, smoothing
 from .tools import collapse_cube
 from .tools.spectral_operations import SpectralOperationHandler
